+++ xarray/tests/test_datatree.py
@@ -72,6 +72,23 @@ class TestNames:
         mary: DataTree = DataTree(children={"Sue": sue})  # noqa
         assert sue.name == "Sue"
 
+    def test_dataset_containing_slashes(self):
+        xda: xr.DataArray = xr.DataArray(
+            [[1, 2]],
+            coords={"label": ["a"], "R30m/y": [30, 60]},
+        )
+        xds: xr.Dataset = xr.Dataset({"group/subgroup/my_variable": xda})
+        with pytest.raises(
+            KeyError,
+            match=re.escape(
+                "Given Dataset contains path-like variable names: "
+                "['R30m/y', 'group/subgroup/my_variable']. "
+                "A Dataset represents a group, and a single group cannot "
+                "have path-like variable names. "
+            ),
+        ):
+            DataTree(xds)
+
 
 class TestPaths:
     def test_path_property(self):